backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
firebase-admin==6.2.0
openai==0.28.1
langchain==0.0.354
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0